refapp-django/requirements.txt
django==3.1
djangorestframework==3.12.2
django-rest-swagger==2.2.0
elasticsearch6==6.8.1
requests==2.25.1
mysql-connector-python
elastic-apm